added audio clip

+++ Assets/Scripts/GameController.cs
@@ -18,7 +18,7 @@ public class GameController : MonoBehaviour
     public GameObject alphabetPrefab;
     
     public Dictionary<int, GameObject> alphabets;
-    public int ROWS = 4;
+    public float ROWS = 4;
     public float WIDTH = 10;
     public int SCORE = 0;
     public float SPAWN_RATE;
